Split data collection and serializing/logging data from event handling:
This is a refactoring to get a cleaner separation between handling
of the ZPublisher events, and the rest of the work that's being done.

Previously everyting was stuffed into subscribers.py, which should
really only be the entry point from which event handlers delegate
to functions that collect data to be logged and serialize and write
it to a logfile.
