+++ ftw/structlog/logger.py
@@ -1,6 +1,7 @@
 from App.config import getConfiguration
 from logging import FileHandler
 from logging import getLogger
+import json
 import logging
 
 
@@ -8,6 +9,10 @@ logger = getLogger('ftw.structlog')
 
 
 def setup_logger():
+    """Set up logger that writes to the JSON based logfile.
+
+    May be invoked multiple times, and must therefore be idempotent.
+    """
     if not logger.handlers:
         path = get_logfile_path()
         handler = FileHandler(path)
@@ -29,6 +34,15 @@ def get_logfile_path():
     assert eventlog_path.endswith('.log')
     path = eventlog_path.replace('.log', '-json.log')
     return path
+
+
+def log_request_data(request_data):
+    """Given a Python dict of key/value pairs to be logged, serialize and log
+    them to a JSON based logfile.
+    """
+    json_log = setup_logger()
+    msg = json.dumps(request_data, sort_keys=True)
+    json_log.info(msg)
 
 
 setup_logger()
