+++ ftw/structlog/subscribers.py
@@ -1,9 +1,8 @@
 from datetime import datetime
-from ftw.structlog.logger import setup_logger
+from ftw.structlog.collector import collect_data_to_log
+from ftw.structlog.logger import log_request_data
 from threading import local
 from tzlocal import get_localzone
-from zope.component.hooks import getSite
-import json
 import logging
 import pytz
 import time
@@ -11,7 +10,6 @@ import time
 
 LOG_TZ = get_localzone()
 
-json_log = setup_logger()
 root_logger = logging.root
 
 
@@ -22,15 +20,21 @@ timing.timestamp = None
 
 
 def handle_pub_start(event):
+    """Handle IPubStart events: Keep track of time when request started.
+    """
     global timing
     # Get time in UTC, make non-naive, and convert to local time for logging
     ts = datetime.utcnow().replace(tzinfo=pytz.utc)
     timing.timestamp = ts.astimezone(LOG_TZ).isoformat()
-
     timing.pub_start = time.time()
 
 
 def handle_pub_end(event):
+    """Handle IPubSuccess or IPubFailure events.
+
+    Log the request, being very careful not to cause any unhandled exceptions
+    as a side effect of logging.
+    """
     try:
         log_request(event)
     except Exception as exc:
@@ -38,58 +42,8 @@ def handle_pub_end(event):
 
 
 def log_request(event):
+    global timing
     request = event.request
 
-    logdata = collect_data_to_log(request)
-    msg = json.dumps(logdata, sort_keys=True)
-    json_log.info(msg)
-
-
-def collect_data_to_log(request):
-    global timing
-    duration = time.time() - timing.pub_start
-    timing.pub_start = None
-
-    logdata = {
-        'host': request.getClientAddr(),
-        'site': get_site_id(),
-        'user': get_username(request),
-        'timestamp': timing.timestamp,
-        'method': request.method,
-        'url': get_url(request),
-        'status': request.response.getStatus(),
-        'bytes': get_content_length(request),
-        'duration': duration,
-        # TODO: Always return empty string if no referrer
-        'referer': request.environ.get('HTTP_REFERER'),
-        'user_agent': request.environ.get('HTTP_USER_AGENT'),
-    }
-
-    return logdata
-
-
-def get_content_length(request):
-    content_length = request.response.getHeader('Content-Length')
-    if content_length:
-        return int(content_length)
-
-
-def get_site_id():
-    site = getSite()
-    if site:
-        return site.id
-    return ''
-
-
-def get_username(request):
-    user = request.get('AUTHENTICATED_USER')
-    if user:
-        return user.getUserName()
-
-
-def get_url(request):
-    url = request.get('ACTUAL_URL')
-    qs = request.get('QUERY_STRING')
-    if qs:
-        url = url + "?" + qs
-    return url
+    request_data = collect_data_to_log(timing, request)
+    log_request_data(request_data)
